server validation for status and removal of multiple records

+++ group_actions/delete.php
@@ -2,8 +2,20 @@
 
 include '../database/connection.php';
 
-//TODO: Add validation
+//validation - user ids must be positive integers separated by comma
+if (preg_match('/^\d+(,\d+)*$/', $_GET["users"]) == false) {
+    echo json_encode([
+        "status" => false,
+        "error" => [
+            "code" => 422,
+            "message" => "User id list contains invalid numbers."
+        ]
+    ]);
 
+    exit;
+}
+
+//Delete multiple users
 $result = executeStatement("DELETE FROM users WHERE id IN ({$_GET["users"]});", []);
 
 if($result === false) {
